% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{avro_deserialize}
\alias{avro_deserialize}
\title{Deserialize Avro data with Confluent wire format}
\usage{
avro_deserialize(client, raw_bytes)
}
\arguments{
\item{client}{A Schema Registry client created with \code{sr_connect}}

\item{raw_bytes}{A raw vector of Confluent wire format Avro data}
}
\value{
An R list containing the deserialized data
}
\description{
Deserialize Avro data with Confluent wire format
}
\examples{
\dontrun{
client <- sr_connect("http://localhost:8081")
record <- avro_deserialize(client, raw_bytes)
}
}
