% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{avro_serialize}
\alias{avro_serialize}
\title{Serialize R data to Avro with Confluent wire format}
\usage{
avro_serialize(client, subject, data)
}
\arguments{
\item{client}{A Schema Registry client created with \code{sr_connect}}

\item{subject}{The subject name (used to look up the schema)}

\item{data}{An R list to serialize}
}
\value{
A raw vector containing the serialized data
}
\description{
Serialize R data to Avro with Confluent wire format
}
\examples{
\dontrun{
client <- sr_connect("http://localhost:8081")
raw_bytes <- avro_serialize(
  client, "my-topic-value", list(id = 1L, name = "test")
)
}
}
