% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{kafka_flush}
\alias{kafka_flush}
\title{Flush the Kafka producer}
\usage{
kafka_flush(producer, timeout = 5000L)
}
\arguments{
\item{producer}{A Kafka producer created with \code{kafka_producer}}

\item{timeout}{Timeout in milliseconds (default 5000)}
}
\description{
Flush the Kafka producer
}
\examples{
\dontrun{
kafka_flush(producer, timeout = 5000)
}
}
