% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{kafka_produce}
\alias{kafka_produce}
\title{Produce a raw byte message to a Kafka topic}
\usage{
kafka_produce(producer, topic, value, key = NULL)
}
\arguments{
\item{producer}{A Kafka producer created with \code{kafka_producer}}

\item{topic}{The Kafka topic name}

\item{value}{A raw vector to send as the message value}

\item{key}{An optional character string key}
}
\description{
Produce a raw byte message to a Kafka topic
}
\examples{
\dontrun{
raw_bytes <- avro_serialize(client, "my-topic-value", data)
kafka_produce(producer, "my-topic", raw_bytes, key = "my-key")
}
}
