% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{kafka_produce_text}
\alias{kafka_produce_text}
\title{Produce a text message to a Kafka topic}
\usage{
kafka_produce_text(producer, topic, value, key = NULL)
}
\arguments{
\item{producer}{A Kafka producer created with \code{kafka_producer}}

\item{topic}{The Kafka topic name}

\item{value}{A character string to send as the message value}

\item{key}{An optional character string key}
}
\description{
Produce a text message to a Kafka topic
}
\examples{
\dontrun{
kafka_produce_text(producer, "my-topic", '{"id": 1}', key = "my-key")
}
}
