% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{kafka_producer}
\alias{kafka_producer}
\title{Create a Kafka producer}
\usage{
kafka_producer(config)
}
\arguments{
\item{config}{A named list of librdkafka configuration values}
}
\value{
An external pointer to the producer object
}
\description{
Create a Kafka producer
}
\examples{
\dontrun{
producer <- kafka_producer(list(
  "bootstrap.servers" = "localhost:9092"
))
}
}
