% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sr_check_compatibility}
\alias{sr_check_compatibility}
\title{Check compatibility of a schema with a subject}
\usage{
sr_check_compatibility(client, subject, schema_json)
}
\arguments{
\item{client}{A Schema Registry client created with \code{sr_connect}}

\item{subject}{The subject name}

\item{schema_json}{The Avro schema as a JSON string}
}
\value{
Logical indicating compatibility
}
\description{
Check compatibility of a schema with a subject
}
\examples{
\dontrun{
client <- sr_connect("http://localhost:8081")
schema <- '{"type":"record","name":"Test",
  "fields":[{"name":"id","type":"int"}]}'
is_compat <- sr_check_compatibility(client, "test-value", schema)
}
}
