% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sr_delete_subject}
\alias{sr_delete_subject}
\title{Delete a subject from the Schema Registry}
\usage{
sr_delete_subject(client, subject)
}
\arguments{
\item{client}{A Schema Registry client created with \code{sr_connect}}

\item{subject}{The subject name}
}
\value{
Logical indicating success
}
\description{
Delete a subject from the Schema Registry
}
\examples{
\dontrun{
client <- sr_connect("http://localhost:8081")
sr_delete_subject(client, "test-value")
}
}
