% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sr_get_schema}
\alias{sr_get_schema}
\title{Get schema JSON for a subject}
\usage{
sr_get_schema(client, subject, version = NULL)
}
\arguments{
\item{client}{A Schema Registry client created with \code{sr_connect}}

\item{subject}{The subject name}

\item{version}{The schema version (NULL for latest)}
}
\value{
A string containing the schema JSON
}
\description{
Get schema JSON for a subject
}
\examples{
\dontrun{
client <- sr_connect("http://localhost:8081")
schema <- sr_get_schema(client, "my-topic-value")
schema_v1 <- sr_get_schema(client, "my-topic-value", version = 1L)
}
}
