% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sr_get_schema_by_id}
\alias{sr_get_schema_by_id}
\title{Get schema JSON by global ID}
\usage{
sr_get_schema_by_id(client, id)
}
\arguments{
\item{client}{A Schema Registry client created with \code{sr_connect}}

\item{id}{The global schema ID (integer, as returned by \code{sr_register_schema})}
}
\value{
A string containing the schema JSON
}
\description{
Get schema JSON by global ID
}
\examples{
\dontrun{
client <- sr_connect("http://localhost:8081")
schema <- sr_get_schema_by_id(client, 1L)
}
}
