% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sr_list_subjects}
\alias{sr_list_subjects}
\title{List all subjects in the Schema Registry}
\usage{
sr_list_subjects(client)
}
\arguments{
\item{client}{A Schema Registry client created with \code{sr_connect}}
}
\value{
A character vector of subject names
}
\description{
List all subjects in the Schema Registry
}
\examples{
\dontrun{
client <- sr_connect("http://localhost:8081")
subjects <- sr_list_subjects(client)
}
}
