% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{sr_register_schema}
\alias{sr_register_schema}
\title{Register a schema under a subject}
\usage{
sr_register_schema(client, subject, schema_json)
}
\arguments{
\item{client}{A Schema Registry client created with \code{sr_connect}}

\item{subject}{The subject name}

\item{schema_json}{The Avro schema as a JSON string}
}
\value{
The schema ID (integer)
}
\description{
Register a schema under a subject
}
\examples{
\dontrun{
client <- sr_connect("http://localhost:8081")
schema <- '{"type":"record","name":"Test","fields":[{"name":"id","type":"int"}]}'
id <- sr_register_schema(client, "test-value", schema)
}
}
